hydrology: make project-river-network physics default
