+++ mods/mod-swooper-maps/src/domain/hydrology/ops/project-river-network/index.ts
@@ -1,9 +1,9 @@
 import { createOp } from "@swooper/mapgen-core/authoring";
 import ProjectRiverNetworkContract from "./contract.js";
-import { defaultStrategy, physicsStrategy } from "./strategies/index.js";
+import { defaultStrategy } from "./strategies/index.js";
 
 const projectRiverNetwork = createOp(ProjectRiverNetworkContract, {
-  strategies: { default: defaultStrategy, physics: physicsStrategy },
+  strategies: { default: defaultStrategy },
 });
 
 export type * from "./types.js";
