+++ mods/mod-swooper-maps/src/domain/hydrology/ops/project-river-network/strategies/index.ts
@@ -1,2 +1,1 @@
 export { defaultStrategy } from "./default.js";
-export { physicsStrategy } from "./physics.js";
