+++ mods/mod-swooper-maps/src/domain/hydrology/shared/knob-multipliers.ts
@@ -109,8 +109,6 @@ export const HYDROLOGY_RIVER_DENSITY_RUNOFF_SCALE_MULTIPLIER = {
 
 export const HYDROLOGY_RIVERS_DEFAULT_MIN_LENGTH = 5 as const;
 export const HYDROLOGY_RIVERS_DEFAULT_MAX_LENGTH = 15 as const;
-export const HYDROLOGY_PROJECT_RIVER_NETWORK_MINOR_PERCENTILE_DEFAULT = 0.85 as const;
-export const HYDROLOGY_PROJECT_RIVER_NETWORK_MAJOR_PERCENTILE_DEFAULT = 0.95 as const;
 
 export const HYDROLOGY_REFINE_RIVER_CORRIDOR_LOWLAND_ADJACENCY_BONUS_BASE = 14 as const;
 export const HYDROLOGY_REFINE_RIVER_CORRIDOR_HIGHLAND_ADJACENCY_BONUS_BASE = 10 as const;
